+++ routes/imageRoute.js
@@ -2,7 +2,7 @@ const express = require("express");
 const router = express.Router();
 const multer = require("multer");
 
-const uploadController = require("../controllers/uploadController");
+const imageController = require("../controllers/imageController");
 
 const storage = multer.diskStorage({
   destination: (req, file, cb) => {
@@ -21,6 +21,8 @@ const upload = multer({
 });
 
 // 파일 업로드
-router.post("/:carId", upload.array("image"), uploadController.uploadImages);
+router.post("/:carId", upload.array("image"), imageController.uploadImages);
+// 파일명으로 이미지 파일 가져오기
+router.get("/:fileName", imageController.getImages);
 
 module.exports = router;
